# A bit faster.