Several minor optimizations. Here are the numbers on my laptop (M2, LLVM 14).

Current main branch
```
$ ./buildmain/benchmarks/bench --benchmark_repetitions=1 --benchmark_filter=Ada
BasicBench_AdaURL      13544 ns        13543 ns        51656 time/byte=16.7615ns
```

The PR split-fragment-state by @anonrig 

```
$ ./build-split-fragment-state/benchmarks/bench --benchmark_repetitions=1 --benchmark_filter=Ada 
BasicBench_AdaURL      11132 ns        11131 ns        63155 time/byte=13.7761ns
```

This PR:
```
$ ./build-dlemire-split-fragment-state/benchmarks/bench --benchmark_repetitions=1 --benchmark_filter=Ada 
BasicBench_AdaURL       9136 ns         9135 ns        76940 time/byte=11.306ns
```

Note: This is a PR on split-fragment-state *not the main branch*. Merging this would modify the branch split-fragment-state.